pyproject.toml
[tool.poetry]
name = "kb-cli"
version = "0.1.0"
description = ""
authors = ["Joseph P. <joseph.pena.q@joseph.pena.q>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.13.0"
pyfzf = "^0.3.1"
pyyaml = "^6.0.2"
pydantic = "^2.9.2"

[tool.poetry.scripts]
kb = "kb_cli.__main__:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
